# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/tournament": {
            "get": {
                "description": "Get all tournaments",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tournament"
                ],
                "summary": "All tournaments",
                "parameters": [
                    {
                        "type": "string",
                        "description": "page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "page size",
                        "name": "count",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "search",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Contest bracket",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dtos.TournamentsResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Failed to return tournament contest",
                        "schema": {
                            "$ref": "#/definitions/dtos.MessageResponseType"
                        }
                    }
                }
            }
        },
        "/tournament/contest/{tournamentId}": {
            "get": {
                "description": "Get tournament contest",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tournament"
                ],
                "summary": "Tournament contest",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tournament id",
                        "name": "tournamentId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "contestType",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Contest bracket",
                        "schema": {
                            "$ref": "#/definitions/dtos.Bracket"
                        }
                    },
                    "400": {
                        "description": "Failed to return tournament contest",
                        "schema": {
                            "$ref": "#/definitions/dtos.MessageResponseType"
                        }
                    }
                }
            }
        },
        "/tournament/create": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Create new tournament for current user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tournament"
                ],
                "summary": "Create new tournament",
                "parameters": [
                    {
                        "description": "Data to create tournament",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.CreateTournament"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tournament created",
                        "schema": {
                            "$ref": "#/definitions/dtos.MessageResponseType"
                        }
                    },
                    "400": {
                        "description": "Error during tournament creation",
                        "schema": {
                            "$ref": "#/definitions/dtos.MessageResponseType"
                        }
                    }
                }
            }
        },
        "/tournament/delete": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Delete tournaments for current user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tournament"
                ],
                "summary": "Delete tournaments",
                "parameters": [
                    {
                        "description": "Data to delete tournaments",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.TournamentIds"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tournaments deleted",
                        "schema": {
                            "$ref": "#/definitions/dtos.MessageResponseType"
                        }
                    },
                    "400": {
                        "description": "Error during tournaments deletion",
                        "schema": {
                            "$ref": "#/definitions/dtos.MessageResponseType"
                        }
                    }
                }
            }
        },
        "/tournament/delete/{tournamentId}": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Delete tournament for current user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tournament"
                ],
                "summary": "Delete tournament",
                "responses": {
                    "200": {
                        "description": "Tournament deleted",
                        "schema": {
                            "$ref": "#/definitions/dtos.MessageResponseType"
                        }
                    },
                    "400": {
                        "description": "Error during tournament deletion",
                        "schema": {
                            "$ref": "#/definitions/dtos.MessageResponseType"
                        }
                    }
                }
            }
        },
        "/tournament/edit/{tournamentId}": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Edit tournament for current user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tournament"
                ],
                "summary": "Edit tournament",
                "parameters": [
                    {
                        "description": "Data to edit tournament",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.EditTournament"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tournament edited",
                        "schema": {
                            "$ref": "#/definitions/dtos.MessageResponseType"
                        }
                    },
                    "400": {
                        "description": "Error during tournament edition",
                        "schema": {
                            "$ref": "#/definitions/dtos.MessageResponseType"
                        }
                    }
                }
            }
        },
        "/tournament/tiktoks/{tournamentId}": {
            "get": {
                "description": "Get tournament tiktoks",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tournament"
                ],
                "summary": "Tournament tiktoks",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tournament id",
                        "name": "tournamentId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tournament tiktoks",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Tiktok"
                            }
                        }
                    },
                    "400": {
                        "description": "Tournament not found",
                        "schema": {
                            "$ref": "#/definitions/dtos.MessageResponseType"
                        }
                    }
                }
            }
        },
        "/tournament/{tournamentId}": {
            "get": {
                "description": "Get tournament details by its id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tournament"
                ],
                "summary": "Tournament details",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tournament id",
                        "name": "tournamentId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tournament",
                        "schema": {
                            "$ref": "#/definitions/models.Tournament"
                        }
                    },
                    "400": {
                        "description": "Tournament not found",
                        "schema": {
                            "$ref": "#/definitions/dtos.MessageResponseType"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Increment wins and increment times_played",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tournament"
                ],
                "summary": "Update tournament winner statistics",
                "parameters": [
                    {
                        "description": "Data to update tournament winner",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.CreateTournament"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Winner updated",
                        "schema": {
                            "$ref": "#/definitions/dtos.MessageResponseType"
                        }
                    },
                    "400": {
                        "description": "Error during winner updating",
                        "schema": {
                            "$ref": "#/definitions/dtos.MessageResponseType"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dtos.Bracket": {
            "type": "object",
            "properties": {
                "countMatches": {
                    "type": "integer"
                },
                "rounds": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dtos.Round"
                    }
                }
            }
        },
...
        "dtos.CreateTiktok": {
            "type": "object",
            "required": [
                "name",
                "url"
            ],
            "properties": {
                "name": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "dtos.CreateTournament": {
            "type": "object",
            "required": [
                "name",
                "photoURL",
                "tiktoks"
            ],
            "properties": {
                "name": {
                    "type": "string"
                },
                "photoURL": {
                    "type": "string"
                },
                "size": {
                    "type": "integer",
                    "maximum": 64,
                    "minimum": 4
                },
                "tiktoks": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dtos.CreateTiktok"
                    }
                }
            }
        },
        "dtos.EditTournament": {
            "type": "object",
            "required": [
                "name",
                "photoURL",
                "tiktoks"
            ],
            "properties": {
                "name": {
                    "type": "string"
                },
                "photoURL": {
                    "type": "string"
                },
                "size": {
                    "type": "integer",
                    "maximum": 64,
                    "minimum": 4
                },
                "tiktoks": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dtos.CreateTiktok"
                    }
                }
            }
        },
        "dtos.Match": {
            "type": "object",
            "properties": {
                "firstOption": {},
                "matchID": {
                    "type": "string"
                },
                "secondOption": {}
            }
        },
        "dtos.MessageResponseType": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "dtos.Round": {
            "type": "object",
            "properties": {
                "matches": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dtos.Match"
                    }
                },
                "round": {
                    "type": "integer"
                }
            }
        },
        "dtos.TournamentIds": {
            "type": "object",
            "required": [
                "tournamentIds"
            ],
            "properties": {
                "tournamentIds": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "dtos.TournamentsResponse": {
            "type": "object",
            "required": [
                "tournamentCount",
                "tournaments"
            ],
            "properties": {
                "tournamentCount": {
                    "type": "integer"
                },
                "tournaments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Tournament"
                    }
                }
            }
        },
...
        "models.Tiktok": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "tournament": {
                    "$ref": "#/definitions/models.Tournament"
                },
                "tournamentID": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                },
                "wins": {
                    "type": "integer"
                }
            }
        },
        "models.Tournament": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "photoURL": {
                    "type": "string"
                },
                "size": {
                    "type": "integer"
                },
                "timesPlayed": {
                    "type": "integer"
                },
                "user": {
                    "$ref": "#/definitions/models.User"
                },
                "userID": {
                    "type": "string"
                }
            }
        },
        "models.User": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "photoURL": {
                    "type": "string"
                }
            }
        }
    }
}