# internal/api/controllers/tournament.controller.go
package controllers

import (
	"fmt"
	"github.com/gofiber/fiber/v2"
	"tiktok-arena/internal/api/controllers/response"
	"tiktok-arena/internal/core/dtos"
	"tiktok-arena/internal/core/services"
	"tiktok-arena/internal/core/validator"
)

type TournamentController struct {
	TournamentService *services.TournamentService
}

func NewTournamentController(tournamentService *services.TournamentService) *TournamentController {
	return &TournamentController{TournamentService: tournamentService}
}

// CreateTournament
//
//	@Summary		Create new tournament
//	@Description	Create new tournament for current user
//	@Tags			tournament
//	@Accept			json
//	@Produce		json
//	@Security		ApiKeyAuth
//	@Param			payload	body		dtos.CreateTournament		true	"Data to create tournament"
//	@Success		200		{object}	dtos.MessageResponseType	"Tournament created"
//	@Failure		400		{object}	dtos.MessageResponseType	"Error during tournament creation"
//	@Router			/tournament/create [post]
func (cr *TournamentController) CreateTournament(c *fiber.Ctx) error {
	user := c.Locals("user")
	userId, err := validator.GetUserIdAndCheckJWT(user)
	if err != nil {
		return err
	}

	var payload *dtos.CreateTournament
	err = c.BodyParser(&payload)
	if err != nil {
		return err
	}

	err = cr.TournamentService.CreateTournament(payload, userId)
	if err != nil {
		return err
	}

	return response.MessageResponse(c, fiber.StatusCreated, fmt.Sprintf("Tournament created %v", payload.Name))
}

// EditTournament
//
//	@Summary		Edit tournament
//	@Description	Edit tournament for current user
//	@Tags			tournament
//	@Accept			json
//	@Produce		json
//	@Security		ApiKeyAuth
//	@Param			payload	body		dtos.EditTournament			true	"Data to edit tournament"
//	@Success		200		{object}	dtos.MessageResponseType	"Tournament edited"
//	@Failure		400		{object}	dtos.MessageResponseType	"Error during tournament edition"
//	@Router			/tournament/edit/{tournamentId} [put]
func (cr *TournamentController) EditTournament(c *fiber.Ctx) error {
	user := c.Locals("user")
	userId, err := validator.GetUserIdAndCheckJWT(user)

	var payload *dtos.EditTournament
	err = c.BodyParser(&payload)
	if err != nil {
		return err
	}

	tournamentIdString := c.Params("tournamentId")
	err = cr.TournamentService.EditTournament(payload, userId, tournamentIdString)
	if err != nil {
		return err
	}

	return response.MessageResponse(c, fiber.StatusOK, fmt.Sprintf("Successfully edited tournament %s", payload.Name))
}

// DeleteTournament
//
//	@Summary		Delete tournament
//	@Description	Delete tournament for current user
//	@Tags			tournament
//	@Accept			json
//	@Produce		json
//	@Security		ApiKeyAuth
//	@Success		200	{object}	dtos.MessageResponseType	"Tournament deleted"
//	@Failure		400	{object}	dtos.MessageResponseType	"Error during tournament deletion"
//	@Router			/tournament/delete/{tournamentId} [delete]
func (cr *TournamentController) DeleteTournament(c *fiber.Ctx) error {
	user := c.Locals("user")
	userId, err := validator.GetUserIdAndCheckJWT(user)

	tournamentIdString := c.Params("tournamentId")
	err = cr.TournamentService.DeleteTournament(userId, tournamentIdString)
	if err != nil {
		return err
	}

	return response.MessageResponse(c, fiber.StatusOK,
		fmt.Sprintf("Successfully deleted tournament %s", tournamentIdString))
}

// DeleteTournaments
//
//	@Summary		Delete tournaments
//	@Description	Delete tournaments for current user
//	@Tags			tournament
//	@Accept			json
//	@Produce		json
//	@Security		ApiKeyAuth
//	@Param			payload	body		dtos.TournamentIds			true	"Data to delete tournaments"
//	@Success		200		{object}	dtos.MessageResponseType	"Tournaments deleted"
//	@Failure		400		{object}	dtos.MessageResponseType	"Error during tournaments deletion"
//	@Router			/tournament/delete [delete]
func (cr *TournamentController) DeleteTournaments(c *fiber.Ctx) error {
	user := c.Locals("user")
	userId, err := validator.GetUserIdAndCheckJWT(user)
	if err != nil {
		return err
	}

	var payload *dtos.TournamentIds
	err = c.BodyParser(&payload)
	if err != nil {
		return err
	}

	err = cr.TournamentService.DeleteTournaments(userId, payload)
	if err != nil {
		return err
	}

	return response.MessageResponse(c, fiber.StatusOK,
		fmt.Sprintf("Successfully deleted tournaments"))
}

// GetAllTournaments
//
//	@Summary		All tournaments
//	@Description	Get all tournaments
//	@Tags			tournament
//	@Accept			json
//	@Produce		json
//	@Param			page				query		string						false	"page number"
//	@Param			count				query		string						false	"page size"
//	@Param			search				query		string						false	"search"
//	@Success		200					{array}		dtos.TournamentsResponse	"Contest bracket"
//	@Failure		400					{object}	dtos.MessageResponseType	"Failed to return tournament contest"
//	@Router			/tournament [get]																																																																																																																																																																																																				[get]
func (cr *TournamentController) GetAllTournaments(c *fiber.Ctx) error {
	q := new(dtos.PaginationQueries)
	if err := c.QueryParser(q); err != nil {
		return err
	}
	dtos.ValidatePaginationQueries(q)
	tournamentResponse, err := cr.TournamentService.GetAllTournaments(q)
	if err != nil {
		return err
	}
	return c.Status(fiber.StatusOK).JSON(tournamentResponse)
}

// GetTournamentDetails
//
//	@Summary		Tournament details
//	@Description	Get tournament details by its id
//	@Tags			tournament
//	@Accept			json
//	@Produce		json
//	@Param			tournamentId	path		string						true	"Tournament id"
//	@Success		200				{object}	models.Tournament			"Tournament"
//	@Failure		400				{object}	dtos.MessageResponseType	"Tournament not found"
//	@Router			/tournament/{tournamentId} [get]
func (cr *TournamentController) GetTournamentDetails(c *fiber.Ctx) error {
	tournamentIdString := c.Params("tournamentId")
	tournament, err := cr.TournamentService.GetTournamentDetails(tournamentIdString)
	if err != nil {
		return err
	}
	return c.Status(fiber.StatusOK).JSON(tournament)
}

// GetTournamentTiktoks
//
//	@Summary		Tournament tiktoks
//	@Description	Get tournament tiktoks
//	@Tags			tournament
//	@Accept			json
//	@Produce		json
//	@Param			tournamentId	path		string						true	"Tournament id"
//	@Success		200				{array}		models.Tiktok				"Tournament tiktoks"
//	@Failure		400				{object}	dtos.MessageResponseType	"Tournament not found"
//	@Router			/tournament/tiktoks/{tournamentId} [get]
func (cr *TournamentController) GetTournamentTiktoks(c *fiber.Ctx) error {
	tournamentIdString := c.Params("tournamentId")
	tiktoks, err := cr.TournamentService.GetTournamentTiktoks(tournamentIdString)
	if err != nil {
		return err
	}
	return c.Status(fiber.StatusOK).JSON(tiktoks)
}

// TournamentWinner
//
//	@Summary		Update tournament winner statistics
//	@Description	Increment wins and increment times_played
//	@Tags			tournament
//	@Accept			json
//	@Produce		json
//	@Security		ApiKeyAuth
//	@Param			payload	body		dtos.CreateTournament		true	"Data to update tournament winner"
//	@Success		200		{object}	dtos.MessageResponseType	"Winner updated"
//	@Failure		400		{object}	dtos.MessageResponseType	"Error during winner updating"
//	@Router			/tournament/{tournamentId} [put]
func (cr *TournamentController) TournamentWinner(c *fiber.Ctx) error {
	user := c.Locals("user")
	_, err := validator.GetUserIdAndCheckJWT(user)

	tournamentIdString := c.Params("tournamentId")

	var payload *dtos.TournamentWinner
	err = c.BodyParser(&payload)
	if err != nil {
		return err
	}

	err = cr.TournamentService.TournamentWinner(tournamentIdString, payload)
	if err != nil {
		return err
	}

	return response.MessageResponse(c, fiber.StatusOK,
		fmt.Sprintf("Successfully registered winner for tournament %s", tournamentIdString))
}

// GetTournamentContest
//
//	@Summary		Tournament contest
//	@Description	Get tournament contest
//	@Tags			tournament
//	@Accept			json
//	@Produce		json
//	@Param			tournamentId	path		string						true	"Tournament id"
//	@Param			payload			query		dtos.ContestPayload			true	"Contest type"
//	@Success		200				{object}	dtos.Bracket				"Contest bracket"
//	@Failure		400				{object}	dtos.MessageResponseType	"Failed to return tournament contest"
//	@Router			/tournament/contest/{tournamentId} [get]
func (cr *TournamentController) GetTournamentContest(c *fiber.Ctx) error {
	tournamentIdString := c.Params("tournamentId")
	contestType := c.Query("type")
	bracket, err := cr.TournamentService.GetTournamentContest(tournamentIdString, contestType)
	if err != nil {
		return err
	}
	return c.Status(fiber.StatusOK).JSON(bracket)
}