# internal/core/dtos/tournament.dto.go
package dtos

import (
	"tiktok-arena/internal/core/models"
)

type TournamentsResponse struct {
	TournamentCount int64               `validate:"required"`
	Tournaments     []models.Tournament `validate:"required"`
}

type TournamentWinner struct {
	TiktokURL string `validate:"required"`
}

type TournamentIds struct {
	TournamentIds []string `validate:"required"`
}

type CreateTournament struct {
	Name     string         `validate:"required"`
	PhotoURL string         `validate:"required"`
	Size     int            `validate:"gte=4,lte=64"`
	Tiktoks  []CreateTiktok `validate:"required"`
}

type EditTournament struct {
	Name     string         `validate:"required"`
	PhotoURL string         `validate:"required"`
	Size     int            `validate:"gte=4,lte=64"`
	Tiktoks  []CreateTiktok `validate:"required"`
}

type Bracket struct {
	CountMatches int
	Rounds       []Round
}

type Round struct {
	Round   int
	Matches []Match
}

type Option interface {
	isOption() bool
}

type MatchOption struct {
	MatchID string
}

func (m MatchOption) isOption() bool {
	return true
}

type TiktokOption struct {
	TiktokURL string
}

func (m TiktokOption) isOption() bool {
	return true
}

type Match struct {
	MatchID      string
	FirstOption  interface{}
	SecondOption interface{}
}

type ContestPayload struct {
	ContestType string `validate:"required"`
}

const (
	SingleElimination = "single_elimination"
	KingOfTheHill     = "king_of_the_hill"
)

func GetAllowedTournamentType() map[string]bool {
	return map[string]bool{
		SingleElimination: true,
		KingOfTheHill:     true,
	}
}

func CheckIfAllowedTournamentType(tournamentType string) bool {
	return GetAllowedTournamentType()[tournamentType]
}